@@ -8,32 +8,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~15.1.4",
+    "@angular/animations": "~15.2.0",
     "@angular/cdk": "^15.0.3",
-    "@angular/common": "~15.1.4",
-    "@angular/compiler": "~15.1.4",
-    "@angular/core": "~15.1.4",
-    "@angular/forms": "~15.1.4",
+    "@angular/common": "~15.2.0",
+    "@angular/compiler": "~15.2.0",
+    "@angular/core": "~15.2.0",
+    "@angular/forms": "~15.2.0",
     "@angular/flex-layout": "^15.0.0-beta.42",
     "@angular/material": "^15.0.3",
-    "@angular/platform-browser": "~15.1.4",
-    "@angular/platform-browser-dynamic": "~15.1.4",
-    "@angular/router": "~15.1.4",
+    "@angular/platform-browser": "~15.2.0",
+    "@angular/platform-browser-dynamic": "~15.2.0",
+    "@angular/router": "~15.2.0",
     "@swimlane/ngx-charts": "^20.1.2",
     "angular-plotly.js": "^4.0.4",
-    "gojs": "^2.3.1",
+    "gojs": "^2.3.3",
     "gojs-angular": "^2.0.6",
-    "plotly.js-dist-min": "^2.18.1",
+    "plotly.js-dist-min": "^2.18.2",
     "rxjs": "~7.8.0",
-    "socket.io": "^4.6.0",
-    "socket.io-client": "^4.6.0",
+    "socket.io": "^4.6.1",
+    "socket.io-client": "^4.6.1",
     "ngx-socket-io": "^4.4.0",
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.1.5",
-    "@angular/cli": "~15.1.5",
-    "@angular/compiler-cli": "~15.1.4",
+    "@angular-devkit/build-angular": "~15.2.0",
+    "@angular/cli": "~15.2.0",
+    "@angular/compiler-cli": "~15.2.0",
     "@types/jasmine": "~4.3.1",
     "@types/plotly.js-dist-min": "^2.3.1",
     "@types/d3": "^7.4.0",
