@@ -24,22 +24,26 @@ export class HomeComponent implements OnInit {
   isAdmin: boolean = false;
   
   ngOnInit(): void {
-    this.userService.getUser().subscribe(
-      user => {
+    this.userService.getUser().subscribe({
+      next: user => {
         if (user.accountType == 'admin')
           this.isAdmin = true;
         else if (user.accountType == 'user')
           this.isUser = true;
         else
           this.confirmDialog.open('unknown account type');
-        this.projectService.getProjects().subscribe(
-          projects => this.projects = projects,
-          err => {
+        this.projectService.getProjects().subscribe({
+          next: projects => this.projects = projects,
+          error: err => {
             if (err.status != 404)
               this.errorAlert.open(err.error);
           }
-        );
+        });
+      },
+      error: err => {
+        if (err.status != 401)
+          this.errorAlert.open(err);
       }
-    );
+    });
   }
 }
